requirements.txt
langchain
langchain-groq
langchain-core
pandas
streamlit
python-dotenv
numpy
PyPDF2
python-docx